refactor: 어색한 페키지명 변경
