+++ app/src/main/java/nextstep/jwp/server/handler/RequestHandler.java
@@ -8,18 +8,18 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.Socket;
 import java.util.Objects;
-import nextstep.jwp.http.header.element.Body;
-import nextstep.jwp.http.header.element.Headers;
+import nextstep.jwp.http.message.element.Body;
+import nextstep.jwp.http.message.element.Headers;
 import nextstep.jwp.server.handler.controller.Controller;
 import nextstep.jwp.server.handler.controller.Controllers;
 import nextstep.jwp.web.presentation.controller.factory.CustomControllerFactory;
 import nextstep.jwp.server.handler.controller.standard.StandardControllerFactory;
 import nextstep.jwp.http.exception.Exceptions;
 import nextstep.jwp.http.exception.InternalServerException;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.request.request_line.RequestLine;
-import nextstep.jwp.http.header.response.HttpResponse;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.request.request_line.RequestLine;
+import nextstep.jwp.http.message.response.HttpResponse;
+import nextstep.jwp.http.message.response.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
