+++ app/src/main/java/nextstep/jwp/server/handler/controller/Controller.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.server.handler.controller;
 
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.response.Response;
 
 public interface Controller {
     Response doService(HttpRequest httpRequest);
