+++ app/src/main/java/nextstep/jwp/server/handler/controller/standard/GetStandardController.java
@@ -1,10 +1,10 @@
 package nextstep.jwp.server.handler.controller.standard;
 
-import nextstep.jwp.http.header.element.HttpStatus;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.request.request_line.HttpMethod;
-import nextstep.jwp.http.header.request.request_line.HttpPath;
-import nextstep.jwp.http.header.response.HttpResponse;
+import nextstep.jwp.http.message.element.HttpStatus;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.request.request_line.HttpMethod;
+import nextstep.jwp.http.message.request.request_line.HttpPath;
+import nextstep.jwp.http.message.response.HttpResponse;
 
 public class GetStandardController implements StandardController {
