+++ app/src/main/java/nextstep/jwp/web/presentation/controller/login/GetLoginController.java
@@ -1,11 +1,11 @@
 package nextstep.jwp.web.presentation.controller.login;
 
-import nextstep.jwp.http.header.element.HttpStatus;
+import nextstep.jwp.http.message.element.HttpStatus;
 import nextstep.jwp.web.presentation.controller.CustomController;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.request.request_line.HttpMethod;
-import nextstep.jwp.http.header.response.HttpResponse;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.request.request_line.HttpMethod;
+import nextstep.jwp.http.message.response.HttpResponse;
+import nextstep.jwp.http.message.response.Response;
 
 public class GetLoginController extends CustomController {
