+++ app/src/main/java/nextstep/jwp/web/presentation/controller/login/PostLoginController.java
@@ -7,10 +7,10 @@ import nextstep.jwp.http.common.ParameterExtractor;
 import nextstep.jwp.web.presentation.controller.CustomController;
 import nextstep.jwp.http.exception.BadRequestException;
 import nextstep.jwp.http.exception.UnauthorizedException;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.request.request_line.HttpMethod;
-import nextstep.jwp.http.header.response.HttpResponse;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.request.request_line.HttpMethod;
+import nextstep.jwp.http.message.response.HttpResponse;
+import nextstep.jwp.http.message.response.Response;
 import nextstep.jwp.web.domain.model.User;
 
 public class PostLoginController extends CustomController {
