+++ app/src/test/java/nextstep/jwp/fixture/Fixture.java
@@ -4,11 +4,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import nextstep.jwp.http.header.element.Body;
-import nextstep.jwp.http.header.element.Headers;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.request.request_line.HttpMethod;
-import nextstep.jwp.http.header.request.request_line.RequestLine;
+import nextstep.jwp.http.message.element.Body;
+import nextstep.jwp.http.message.element.Headers;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.request.request_line.HttpMethod;
+import nextstep.jwp.http.message.request.request_line.RequestLine;
 
 public class Fixture {
     private static final String lineSeparator = "\r\n";
