+++ app/src/test/java/nextstep/jwp/http/BodyTest.java
@@ -3,7 +3,7 @@ package nextstep.jwp.http;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import nextstep.jwp.http.common.PathUtils;
-import nextstep.jwp.http.header.element.Body;
+import nextstep.jwp.http.message.element.Body;
 import org.junit.jupiter.api.Test;
 
 class BodyTest {
