+++ app/src/test/java/nextstep/jwp/http/ContentTypeTest.java
@@ -2,7 +2,7 @@ package nextstep.jwp.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import nextstep.jwp.http.header.element.ContentType;
+import nextstep.jwp.http.message.element.ContentType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
