+++ app/src/test/java/nextstep/jwp/http/HttpVersionTest.java
@@ -3,7 +3,7 @@ package nextstep.jwp.http;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.jwp.http.header.element.HttpVersion;
+import nextstep.jwp.http.message.element.HttpVersion;
 import org.junit.jupiter.api.Test;
 
 class HttpVersionTest {
