+++ app/src/test/java/nextstep/jwp/http/message/request/HeadersTest.java
@@ -1,8 +1,8 @@
-package nextstep.jwp.http.header.request;
+package nextstep.jwp.http.message.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import nextstep.jwp.http.header.element.Headers;
+import nextstep.jwp.http.message.element.Headers;
 import org.junit.jupiter.api.Test;
 
 class HeadersTest {
