+++ app/src/test/java/nextstep/jwp/http/message/request/request_line/RequestLineTest.java
@@ -1,9 +1,9 @@
-package nextstep.jwp.http.header.request.request_line;
+package nextstep.jwp.http.message.request.request_line;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import nextstep.jwp.fixture.Fixture;
-import nextstep.jwp.http.header.element.HttpVersion;
+import nextstep.jwp.http.message.element.HttpVersion;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
