+++ app/src/test/java/nextstep/jwp/web/presentation/controller/login/GetLoginControllerTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
 import nextstep.jwp.fixture.Fixture;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
