+++ app/src/test/java/nextstep/jwp/web/presentation/controller/login/GetRegisterControllerTest.java
@@ -3,8 +3,8 @@ package nextstep.jwp.web.presentation.controller.login;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import nextstep.jwp.fixture.Fixture;
-import nextstep.jwp.http.header.request.HttpRequest;
-import nextstep.jwp.http.header.response.Response;
+import nextstep.jwp.http.message.request.HttpRequest;
+import nextstep.jwp.http.message.response.Response;
 import org.junit.jupiter.api.Test;
 
 class GetRegisterControllerTest {
